refactor: Required Type RowMapper 클래스 분리
