+++ jdbc/src/main/java/org/springframework/jdbc/core/statementexecutor/StatementExecutor.java
@@ -1,6 +1,6 @@
 package org.springframework.jdbc.core.statementexecutor;
 
-import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.rowmapper.RowMapper;
 
 import java.sql.PreparedStatement;
